Split mouse_controller from usb_report_updater.
