+++ right/src/usb_report_updater.h
@@ -8,8 +8,6 @@
 
 // Macros:
 
-    #define ACTIVE_MOUSE_STATES_COUNT (SerializedMouseAction_Last + 1)
-
     #define IS_SECONDARY_ROLE_MODIFIER(secondaryRole) (SecondaryRole_LeftCtrl <= (secondaryRole) && (secondaryRole) <= SecondaryRole_RightSuper)
     #define IS_SECONDARY_ROLE_LAYER_SWITCHER(secondaryRole) (SecondaryRole_Mod <= (secondaryRole) && (secondaryRole) <= SecondaryRole_Mouse)
     #define SECONDARY_ROLE_MODIFIER_TO_HID_MODIFIER(secondaryRoleModifier) (1 << ((secondaryRoleModifier) - 1))
@@ -33,40 +31,8 @@
         SecondaryRole_Mouse
     } secondary_role_t;
 
-    typedef enum {
-        MouseSpeed_Normal,
-        MouseSpeed_Accelerated,
-        MouseSpeed_Decelerated,
-    } mouse_speed_t;
-
-    typedef struct {
-        bool isScroll;
-        bool wasMoveAction;
-        serialized_mouse_action_t upState;
-        serialized_mouse_action_t downState;
-        serialized_mouse_action_t leftState;
-        serialized_mouse_action_t rightState;
-        mouse_speed_t prevMouseSpeed;
-        float intMultiplier;
-        float currentSpeed;
-        float targetSpeed;
-        uint8_t initialSpeed;
-        uint8_t acceleration;
-        uint8_t deceleratedSpeed;
-        uint8_t baseSpeed;
-        uint8_t acceleratedSpeed;
-        float xSum;
-        float ySum;
-        int16_t xOut;
-        int16_t yOut;
-        int8_t verticalStateSign;
-        int8_t horizontalStateSign;
-    } mouse_kinetic_state_t;
-
 // Variables:
 
-    extern mouse_kinetic_state_t MouseMoveState;
-    extern mouse_kinetic_state_t MouseScrollState;
     extern uint32_t UsbReportUpdateCounter;
     extern volatile uint8_t UsbReportUpdateSemaphore;
     extern bool TestUsbStack;
